Created pluggable table architecture
Prior to this commit, much of the code was tightly integrated with the physical
storage format of the tables.  This is now gone in favor of a simple interface
that can be reimplemented with different storage characteristics, for example
to allow modifying an existing table, adding a WAL and MVCC.  Emphasis has been
put on avoiding copying data.  I anticipate the interface will undergo some
more as support for more expressive queries are implemented.

The ts-correlate program was disabled from compilation, soon to be replaced by
functionality in ca-shell.

Data type indicators have been removed from the tables themselves; the plan is
to enforce these constraints through the schema table.

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>