schema.c: Add the schema table itself to any newly created schema table
This is required by Codd's rule 4, and effectively allows inspecting the schema
from ca-shell.  An absolute path is required for schema creation, to increase
the probability of successful later access to the schema.

The continued use of a special TABLE_DECLARATION type is unfortunate, and
should be fixed when the data manipulation API becomes stronger.

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>