Started on the ca_schema_query() function and related shell syntax
This will be a highly specialized function for performing set algebraic queries
on inverted file indexes, and returning file summaries.

The query syntax is for now:

  QUERY [query string] WITH (INDEX = [index table], SUMMARY = [summary table]);

The index table is required to be a TEXT -> SORTED_UINT mapping.  The integers
will be treated as offsets in the summary table.

The summary table is required to contain TEXT values in its second column.
Recall that the first columns is always the primary key with the current
architecture.

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>