Added tests for the "write-once" backend
I got nervous about breaking stuff after tonights refactoring, so it was time
to add a test.  The test adds a bunch of values, sorts them and verifies that
they can be read sequentially and in the original order.

This test was also used to discover the binary search bug fixed in 4dd89a8.

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>