Added the "flex" table backend, with MVCC support
Thanks to MVCC, this backend supports DELETE operations.  Combined, with
INSERT, this allows UPDATE to be implemented as well, although at a higher
abstraction level.

Furthermore, this is the first backend that cannot be concurrently written to
by multiple processes.  For example, inserting into the tables performs an
lseek followed by a write.  The plan is to use explicit locking to prevent
corruption.

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>