Improve support for formatting the SELECT output expression list
The architecture to accomplish this is quite convoluted at the moment.  There's
a large amount of branching to deal with different expression types and value
types.  When we move to LLVM based query processing, the code that's actually
being run per row will be significantly simple, as the branching will happen
once per query, rather than once per value.

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>