Use LLVM to compile constant expressions to native code
Only constant expressions are supported for now.  The next steps are adding
support for fields and basic comparison operators, which will greatly increase
the usefulness of this code.

Currently, generated code is never freed, so running a large number of queries
in a single session will use a lot of memory.  This is not a concern with
current use, which is basically limited to "SELECT ... WHERE TRUE" and "SELECT
... WHERE FALSE".

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>