schema.c: Avoid opening ca_catalog.ca_tables twice
We open the table once to read table names.  There's no need to open it again
when SHOW TABLES; is executed.

I noticed a problem while working on this; if the ca_catalog tables are
modified with UPDATE, they are not refreshed.  In the future, CREATE TABLE and
ALTER TABLE operations should map directly to INSERT and UPDATE statements, and
the in-memory copy should be updated accordingly.

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>