Support simple UPDATE statements
To prevent dirty reads, changes are written to new files.  When the transaction
is committed, a new ca_catalog.ca_tables is written with the path names of the
new table files.  Because replacing ca_catalog.ca_tables is atomic, each client
will only see a snapshot.  Before the database can be considered ACID, the
following items must be attended to:

TODO: Allocate a global XID at the first moment data is being accessed within a
      transaction

TODO: Refresh schema cache when a XID is allocated

TODO: Remove old table files when the highest running XID is higher than the
      XID that removed the table file from the schema

TODO: Verify that a table is locked before updating it

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>