Reload schema at the first data statement in a transaction
If the schema is not reloaded, we risk working on a stale copy.  Locking will
ensure that the schema doesn't become stale while we're working on it.

Due to how frequently this operation must be performed, we should later add
optimizations to ensure a schema isn't reloaded when it hasn't been changed,
and make the parsing of the schema as fast as it can possibly be.

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>