Document the schema format planned for version 1.0
The current plan for version 1.0 is to remove functionality that's outside the
short-term needs of dailygieselmann.com, and make sure the tools are close to
optimal for those needs.  This means limiting ourselves to summary tables,
inverted index tables and time series tables.  The query language will be
reduced and specialized, the LLVM dependency and transaction logic will go
away.

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>