writeonce: Use hashing instead of binary search
So far we've been using binary search to look up items in writeonce tables.
However, some of the inverted index tables have tens of millions of rows,
resulting in more than 20 seeks per lookup.  This commit replaces the sorted
array of offsets with a hash table with a load factor of 50%.  Furthermore, we
allow offset values to be either 16, 32 or 64 bits.

We use thte X31 hash[1], an open addressing scheme and fibonacci chaining.

Also, lookups into summary tables have required an access to the sorted array
of offsets in order to find the size of the row.  After this change, row sizes
are store inline in the table, so that we can seek directly to the desired
position.

The one major optimization we can do beyond this commit, is storing inverted
index keys in hash order, so that colliding entries are guaranteed to be stored
physically adjacent, eliminating any extranous seeks during lookup.

1. https://mail.gnome.org/archives/gtk-devel-list/2000-February/msg00057.html

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>