Added support for "summary-override" tables
These tables are to be used when the summary for a document changes, and we
want the changes to be reflected immediately without rebuilding the entire
index.  Utilizing these will have a cost on every single result returned, in
the form of a hashmap lookup, so they should be used sparingly.

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>