schema: Support global prefixes for time-series tables
If you have many items needing the same prefix, this commit eliminates the need
to store that prefix in the table itself.  For example, we have millions of
Wikipedia articles we want to prefix with "wikipedia.en:" for search purposes.

This change is also an optimization, in that table files with non-matching
prefixes need not be scanned when searching for a key with a different prefix.

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>